chatbot/requirements.txt
langchain
langchain_community
python-dotenv
langchain_google_genai
pypdf
flask
langchain-astradb
datasets
ipykernel

-e .